verifactai/requirements.txt
# VeriFactAI — Production Dependencies
# Core ML/NLP
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
spacy>=3.7.0
datasets>=2.15.0

# LLM APIs
anthropic>=0.39.0
openai>=1.6.0

# Web UI
streamlit>=1.29.0
plotly>=5.18.0
pywebview>=5.1

# Data processing
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
